epanet/requirements.txt
Django==3.0.7
asgiref==3.2.7
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
scapy==2.4.3
psycopg2-binary==2.8.5
drf-yasg==1.17.1
django-crispy-forms==1.9.0
django-email-bandit==1.5
django-python3-ldap==0.11.2
django-rest-swagger==2.2.0
djangorestframework-api-key==2.0.0
influxdb==5.3.1
wntr==0.3.0
pyyaml==5.4.1
paramiko==2.7.2
scp==0.13.3
sphinx==4.1.2
sphinx_rtd_theme==0.5.2

